+++ src/types/Albums.js
@@ -5,23 +5,23 @@ const {
 
 module.exports = new GraphQLObjectType({
     name: 'Albums',
-    description: 'Album Attributes',
+    description: 'Album Attributes.',
     fields: () => ({
         id: {
             type: GraphQLString,
-            description: 'Album unique identifier'
+            description: 'Album unique identifier.'
         },
         name: {
             type: GraphQLString,
-            description: 'Album name'
+            description: 'Album name.'
         },
         type: {
             type: GraphQLString,
-            description: 'Album type'
+            description: 'Album type.'
         },
         year: {
             type: GraphQLString,
-            description: 'Release year'
+            description: 'Release year.'
         }
     })
 });