+++ src/types/Band.js
@@ -14,75 +14,75 @@ const { getBandDiscs, getBandReviews } = require('../resolvers');
 
 module.exports = new GraphQLObjectType({
     name: 'Band',
-    description: 'Band Attributes',
+    description: 'Band Attributes.',
     fields: () => ({
         id: {
             type: GraphQLString,
-            description: 'Band unique identifier'
+            description: 'Band unique identifier.'
         },
         name: {
             type: GraphQLString,
-            description: 'Band name'
+            description: 'Band name.'
         },
         genre: {
             type: GraphQLString,
-            description: 'Band genres'
+            description: 'Band genres.'
         },
         country: {
             type: GraphQLString,
-            description: 'Country of origin'
+            description: 'Country of origin.'
         },
         location: {
             type: GraphQLString,
-            description: 'Origin location'
+            description: 'Origin location.'
         },
         themes: {
             type: GraphQLString,
-            description: 'Lyrics themes'
+            description: 'Lyrics themes.'
         },
         status: {
             type: GraphQLString,
-            description: 'Band status'
+            description: 'Band status.'
         },
         label: {
             type: GraphQLString,
-            description: 'Current label'
+            description: 'Current label.'
         },
         formYear: {
             type: GraphQLString,
-            description: 'Formation year'
+            description: 'Formation year.'
         },
         yearsActive: {
             type: GraphQLString,
-            description: 'Years active'
+            description: 'Years active.'
         },
         photoUrl: {
             type: GraphQLString,
-            description: 'Band photo url'
+            description: 'Band photo url.'
         },
         logoUrl: {
             type: GraphQLString,
-            description: 'Band logo url'
+            description: 'Band logo url.'
         },
         discography: {
             type: GraphQLList(AlbumsType),
-            description: 'Band discography',
+            description: 'Band discography.',
             args: {
                 type: {
                     type: GraphQLString,
-                    description: 'Albums type. Can be (`null` or `all`), `main`, `lives`, `demos` or `misc`'
+                    description: 'Albums type. [`all` | `main` | `lives` | `demos` | `misc`].'
                 }
             },
             resolve: async (band, args) => await getBandDiscs({ id: band.id, type: args.type })
         },
         reviews: {
             type: GraphQLList(ReviewsType),
-            description: `Band reviews\\
-                (Always show 200 records)`,
+            description: `Band reviews.\\
+                (Always show 200 records).`,
             args: {
                 start: {
                     type: GraphQLInt,
-                    description: 'Specifies start index to show reviews'
+                    description: 'Specifies start index to show reviews.'
                 }
             },
             resolve: async (band, args) => await getBandReviews({ id: band.id, start: args.start })