+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2019-02-24
+### Added
+- Reviews Query (Get all reviews).
+
+### Fixed
+- Fixed wrong date when get Review
+
 ## [0.0.2] - 2019-01-25
 ### Added
 - Review Query.